refactor: refactor redux integration 3
