+++ src/store/store.ts
@@ -1,8 +1,8 @@
 import { configureStore } from '@reduxjs/toolkit';
-import { TypedUseSelectorHook, useDispatch, useSelector } from 'react-redux';
-import { productsSlice } from './reducers/productsSlice';
-import { productApi } from './services/api';
-import { productDetailsSlice } from './reducers/productDetailsSlice';
+import { TypedUseSelectorHook, useSelector } from 'react-redux';
+import { productApi } from '../services/api';
+import { productsSlice } from './slices/products.slice';
+import { productDetailsSlice } from './slices/productDetails.slice';
 
 function setupStore(initialState = {}) {
   return configureStore({
@@ -20,8 +20,5 @@ function setupStore(initialState = {}) {
 export const store = setupStore();
 
 export type RootState = ReturnType<typeof store.getState>;
-export type AppStore = ReturnType<typeof setupStore>;
-export type AppDispatch = AppStore['dispatch'];
 
-export const useAppDispatch: () => AppDispatch = useDispatch;
 export const useAppSelector: TypedUseSelectorHook<RootState> = useSelector;
